test: 컨트롤러 클래스명으로 변경으로 인한 테스트 클래스명 변경
